relocate calendar components into subdirectories and move all components to main calendar screen directory
