+++ src/screens/Calendar/MeetingModal/styles/DisplayMeetingModalStyles.scss
