+++ src/screens/Calendar/MeetingModal/styles/MeetingModalStyles.scss
