+++ src/screens/Calendar/MeetingModal/styles/MemberCheckbox.scss
