+++ src/screens/Calendar/MeetingModal/styles/SelectAttendees.scss
