+++ src/screens/Calendar/index.ts
@@ -1,3 +1,3 @@
 export * from './CalendarHeader'
 export * from './CalendarTabs'
-export * from './MeetingModal/MeetingModal'
+export * from './MeetingModal'
